+++ CHANGES.rst
@@ -92,6 +92,7 @@ Changelog
 
 **Fixed**
 
+- #1214 Disallow entry of analysis results if the sample is not yet received
 - #1213 Fix instrument notification display in Manage Results View
 - #1212 Fix typo in SamplingFieldsVisibility
 - #1191 Some worksheets pre-1.3 with published analyses remain in open/to_be_verified state